+++ lib/src/emoji_data.dart
@@ -1,0 +1,11 @@
+class EmojiData {
+  final String id;
+  final String name;
+  final String unified;
+  final String char;
+  final String category;
+  final int skin;
+
+  const EmojiData(
+      this.id, this.name, this.unified, this.char, this.category, this.skin);
+}